\name{WECCA-package}
\alias{WECCA-package}
\alias{WECCA}
\docType{package}
\title{
Weighted clustering of array CGH data
}
\description{
It clusters hierarchically the samples on the basis of their DNA copy number profile. 
}
\author{ Wessel N. van Wieringen: \email{w.vanwieringen@vumc.nl} }
\references{ 
Van Wieringen, W.N., Van de Wiel, M.A., Ylstra, B. (2008), "Weighted clustering of called aCGH data", \emph{Biostatistics}, 9(3), 484-500.
}

\keyword{ package }
